Add directorySensitivity to the strategy hash
for the fingerprinting strategies which actually use it.
