+++ subprojects/normalization-java/src/main/java/org/gradle/internal/fingerprint/classpath/impl/ClasspathFingerprintingStrategy.java
@@ -31,7 +31,6 @@ import org.gradle.api.internal.changedetection.state.RuntimeClasspathResourceHas
 import org.gradle.api.internal.changedetection.state.ZipHasher;
 import org.gradle.internal.RelativePathSupplier;
 import org.gradle.internal.file.FileType;
-import org.gradle.internal.fingerprint.DirectorySensitivity;
 import org.gradle.internal.fingerprint.FileSystemLocationFingerprint;
 import org.gradle.internal.fingerprint.FingerprintHashingStrategy;
 import org.gradle.internal.fingerprint.LineEndingSensitivity;
@@ -89,7 +88,7 @@ public class ClasspathFingerprintingStrategy extends AbstractFingerprintingStrat
         ResourceSnapshotterCacheService cacheService,
         Interner<String> stringInterner
     ) {
-        super(identifier, DirectorySensitivity.DEFAULT, zipHasher);
+        super(identifier, zipHasher);
         this.nonZipFingerprintingStrategy = nonZipFingerprintingStrategy;
         this.classpathResourceHasher = classpathResourceHasher;
         this.cacheService = cacheService;
